- DirectXInput: Mouse sensitivity can now set to double values for improved accuracy.

+++ DirectXInput/Keyboard/ControllerHandlers.cs
@@ -27,7 +27,7 @@ namespace DirectXInput.KeyboardCode
                 if (GetSystemTicksMs() >= vControllerDelay_Mouse)
                 {
                     //Get the mouse move amount
-                    int moveSensitivity = Convert.ToInt32(Setting_Load(vConfigurationDirectXInput, "KeyboardMouseMoveSensitivity"));
+                    double moveSensitivity = Convert.ToDouble(Setting_Load(vConfigurationDirectXInput, "KeyboardMouseMoveSensitivity"));
                     GetMouseMovementAmountFromThumb(moveSensitivity, ControllerInput.ThumbLeftX, ControllerInput.ThumbLeftY, true, out int moveHorizontalLeft, out int moveVerticalLeft);
 
                     //Move the mouse cursor
